+++ examples/data-sources/nxos_ipv4_interface/data-source.tf
@@ -1,0 +1,4 @@
+data "nxos_ipv4_interface" "example" {
+  vrf          = "default"
+  interface_id = "eth1/59"
+}