+++ examples/data-sources/nxos_ipv4_interface_address/data-source.tf
@@ -1,0 +1,5 @@
+data "nxos_ipv4_interface_address" "example" {
+  vrf          = "default"
+  interface_id = "eth1/59"
+  address      = "1.1.1.1/24"
+}