+++ examples/resources/nxos_ipv4_interface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ipv4_interface.example "sys/ipv4/inst/dom-[default]/if-[eth1/59]"