+++ examples/resources/nxos_ipv4_interface_address/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ipv4_interface_address.example "sys/ipv4/inst/dom-[default]/if-[eth1/59]/addr-[1.1.1.1/24]"