+++ gen/templates/import.sh
@@ -1,7 +1,1 @@
-{{- $name := .Name }}
-{{- $dn := .Dn }}
-{{- range  .Attributes}}
-{{- if eq .Id true -}}
-terraform import nxos_{{snakeCase $name}}.example "{{sprintf $dn .Example}}"
-{{- end}}
-{{- end}}
+terraform import nxos_{{snakeCase .Name}}.example "{{getExampleDn .Dn .Attributes}}"