+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.2 (unreleased)
+
+- Add nxos_ipv4_interface resource and data source
+- Add nxos_ipv4_interface_address resource and data source
+
 ## 0.2.1
 
 - Add nxos_pyhsical_interface resource and data source